requirements.txt
requests~=2.28.1
selenium
pandas~=1.4.4
beautifulsoup4~=4.11.1
tqdm~=4.64.1
webdriver-manager==3.8.5
psycopg2-binary==2.9.5
sqlalchemy~=1.4.39
loguru==0.6.0